Text update optimizations & reliability (#1)
* batch uploads to Supabase

* sketch server reconciliation

* initial draft

* small fixes

* batch reducer

* use for editor

* bug fix

* fix query

* send initial state

* rewrite state using tr

* get load working
For some reason, useEditor returns a different editor instance the second render. We need to use its initialState to match schemas, so we need to make it a hook dependency and trust callers to useMemo it.

* fast typing fix attempt

* working on delete-all bug

* reserve space in ElementIdGenerator properly

* remove logs

* proofread PR

* remove delete-all hack
The step is actually fine. The bug was caused by the ElementIdGenerator bulk issue fixed afterwards.

* change diff query to avoid duplicate rows
I think if it was inserted and then updated before onChange ran, there would be two rows in the DIFF table, hence JOIN was giving us two output rows.

* isActive toggle for testing concurrency

* maybeMaybeStep

* throttle todo